RemoteControl/requirements.txt
flask>=2.0.0
flask-sock>=0.6.0
mss>=9.0.0
pyautogui>=0.9.54
pillow>=10.0.0
